Address Linux packaging review feedback
