+++ packages/desktop/electron/main/sidecar.ts
@@ -4,8 +4,11 @@ import fs from "node:fs";
 import net from "node:net";
 import path from "node:path";
 import type { Readable } from "node:stream";
-import { resolveDatabasePath } from "./app-paths";
+import { resolveConfigDir, resolveDatabasePath } from "./app-paths";
+import { ensureLinuxSidecarExecutable } from "./sidecar-executable";
 import { NEWER_DATABASE_RECOVERY_DETAIL, isNewerDatabaseStartupFailure } from "./startup-recovery";
+
+export { ensureLinuxSidecarExecutable } from "./sidecar-executable";
 
 const SIDECAR_PORT = 5004;
 const HEALTH_RETRIES = 60;
@@ -192,53 +195,30 @@ export function serviceEnv(
   };
 }
 
-export function ensureLinuxSidecarExecutable(binary: string, platform: NodeJS.Platform): void {
-  if (platform !== "linux") return;
-
-  // Package-managed installs may be root-owned; avoid chmod if +x is present.
-  if (isExecutable(binary)) return;
-
-  try {
-    fs.chmodSync(binary, 0o755);
-  } catch (error) {
-    // Read-only packages can reject chmod even when the binary is executable.
-    const code = errnoCode(error);
-    if ((code === "EPERM" || code === "EROFS" || code === "EACCES") && isExecutable(binary)) {
-      return;
-    }
-    const message = error instanceof Error ? error.message : String(error);
-    throw new Error(`Failed to mark cadencr-service executable at ${binary}: ${message}`);
-  }
-}
-
-function isExecutable(binary: string): boolean {
-  try {
-    fs.accessSync(binary, fs.constants.X_OK);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-function errnoCode(error: unknown): string | undefined {
-  if (typeof error !== "object" || error === null) return undefined;
-  const code = (error as { code?: unknown }).code;
-  return typeof code === "string" ? code : undefined;
-}
-
 export function serviceArgs(
   dbPath: string,
   appVersion?: string,
   rendererDir?: string | null,
 ): string[] {
-  // Settings live beside the database; the service creates the dir on startup.
-  const settingsDir = path.join(path.dirname(path.dirname(dbPath)), "settings");
+  const settingsDir = resolveSettingsDir(dbPath);
 
   const args = ["--db-path", dbPath, "--settings-dir", settingsDir, "--port", String(SIDECAR_PORT)];
   if (appVersion) args.push("--app-version", appVersion);
   // Enables remote-access serving; the local window still loads file://.
   if (rendererDir) args.push("--renderer-dir", rendererDir);
   return args;
+}
+
+export function resolveSettingsDir(
+  dbPath: string,
+  platform: NodeJS.Platform = process.platform,
+  configDir?: string,
+): string {
+  if (platform === "linux") {
+    return path.join(configDir ?? resolveConfigDir(), "settings");
+  }
+  // Preserve the legacy macOS layout beside ~/.cadencr/database.
+  return path.join(path.dirname(path.dirname(dbPath)), "settings");
 }
 
 export function describeStartupFailure(
